Cleans up
