+++ package.json
@@ -74,6 +74,7 @@
     "express-spam-referral-blocker": "^1.2.1",
     "express-subdomain": "^1.0.4",
     "file-loader": "^1.1.11",
+    "filesize": "^4.1.2",
     "graphql": "^14.1.1",
     "graphql-anywhere": "^4.1.3",
     "graphql-tag": "^2.10.1",
