--- react/util/formatFileSize.js
@@ -1,18 +1,0 @@
-const SUFIXES = [
-  'Bytes',
-  'KB',
-  'MB',
-  'GB',
-  'TB',
-  'PB',
-  'EB',
-  'ZB',
-  'YB',
-];
-
-export default (bytes) => {
-  if (!bytes) return '0 Bytes';
-  const i = Math.floor(Math.log(bytes) / Math.log(1024));
-  return `${(bytes / (1024 ** i)).toFixed(2)} ${SUFIXES[i]}`;
-};
-
